Support Nexus Repository Manager as an asset host
Add configuration to NTEL and Ptxdist builds to use our asset host,
Nexus, as a source for downloads and a destination for uploads.

Python distutils will use Nexus as a PyPI source.
PTXdist builds will use Nexus as caching mirrors for several sites.
Private dependencies are stored on Nexus.
Orphaned (unavailable for download) dependencies are stored on Nexus.
Archives of successful builds will be uploaded to Nexus.